text for points and time done

+++ Assets/Scripts/PlayerData.cs
@@ -4,11 +4,12 @@ using UnityEngine;
 
 public class PlayerData : MonoBehaviour
 {
-	public int	level;
-	public int	apples;
-	public int	oranges;
-	public int	bananas;
-	public int	totalFruits;
-	public int	points;
-	public int	time;
+	public int		level;
+	public int		apples;
+	public int		oranges;
+	public int		bananas;
+	public int		totalFruits;
+	public int		points;
+	public float	time;
 }
+
